add manger role

+++ Controllers/productController.js
@@ -3,13 +3,9 @@ const Product = require('../models/productModel');
 const uploadMedia = require('../utils/uploadMedia');
 const ApiFeatures = require('../utils/apiFeatures');
 const AppError = require('../utils/appError');
-const Category = require('../models/categoryModel');
-const User = require('../models/userModel');
 
 exports.createProduct = expressAsync(async (req, res, next) => {
-  const user = User.findById(req.user.id);
-
-  if (user.role === 'admin') {
+  if (req.user.role === 'admin') {
     const uploadRes = await uploadMedia(req.file.path);
     const { name, price, color, quantity, description, category } = req.body;
     if (uploadRes) {
@@ -27,7 +23,7 @@ exports.createProduct = expressAsync(async (req, res, next) => {
         data: product,
       });
     }
-  } else if (user.role === 'manger') {
+  } else if (req.user.role === 'manger') {
     const uploadRes = await uploadMedia(req.file.path);
     const { name, price, color, quantity, description, category, user } =
       req.body;
@@ -85,14 +81,26 @@ exports.getOneProduct = expressAsync(async (req, res, next) => {
 });
 
 exports.updateOneProduct = expressAsync(async (req, res, next) => {
-  const product = await Product.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  let product;
+  product = await Product.findById(req.params.id);
+
+  if (req.user.id !== product.user.id) {
+    return next(
+      new AppError(
+        'You Do Not Have Permission To Perform This Action Only Manger of This Product And Admin',
+        403
+      )
+    );
+  }
 
   if (!product) {
     return next(new AppError(`NO Product Found With Id ${req.params.id}`));
   }
+
+  product = await Product.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true,
+  });
 
   res.status(200).json({
     status: 'success',
@@ -101,11 +109,24 @@ exports.updateOneProduct = expressAsync(async (req, res, next) => {
 });
 
 exports.deleteOneProduct = expressAsync(async (req, res, next) => {
-  const product = await Product.findByIdAndDelete(req.params.id);
+  let product;
+
+  product = await Product.findById(req.params.id);
 
   if (!product) {
     return next(new AppError(`NO Product Found With Id ${req.params.id}`));
   }
+
+  if (req.user.id !== product.user.id) {
+    return next(
+      new AppError(
+        'You Do Not Have Permission To Perform This Action Only Manger of This Product And Admin',
+        403
+      )
+    );
+  }
+
+  product = await Product.findByIdAndDelete(req.params.id);
 
   res.status(204).json({
     status: 'success',
